# tanks_backend/src/main/resources/documentation/openAPI.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Tanks Online API",
    "description": "API documentation for ATA systems multiplayer game",
    "contact": {
      "name": "Antitalent Team",
      "email": "omazicantonio7@gmail.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/api/game-sessions/{sessionId}/events": {
      "get": {
        "tags": [
          "Game"
        ],
        "summary": "Get game events",
        "description": "Retrieves all events for a specific game session",
        "operationId": "getGameEvents",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "description": "ID of the game session",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Game session not found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GameEvent"
                  }
                }
              }
            }
          },
          "200": {
            "description": "Events retrieved successfully",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GameEvent"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Game"
        ],
        "summary": "Create game event",
        "description": "Records a new game event for a specified session",
        "operationId": "createGameEvent",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "description": "ID of the game session",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GameEvent"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Invalid event data",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GameEvent"
                }
              }
            }
          },
          "201": {
            "description": "Event created successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GameEvent"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GameEvent": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "gameSessionId": {
            "type": "integer",
            "format": "int64"
          },
          "eventType": {
            "type": "string",
            "enum": [
              "KILL",
              "DEATH",
              "WIN",
              "LOSE",
              "JOIN",
              "LEAVE"
            ]
          },
          "eventTime": {
            "type": "string",
            "format": "date-time"
          },
          "event_summary": {
            "type": "string"
          }
        }
      },
...
    }
  }
}